+++ src/services/mails/mails.ts
@@ -21,8 +21,7 @@ export default (client: HttpClient, httpAuth: HttpInstance): MailsService => ({
 
   async findById(this: MailsService, id, options) {
     const rqlWithId = rqlBuilder(options?.rql).eq('id', id).build();
-    const res = await this.find({ ...options, rql: rqlWithId });
-    return res.data[0];
+    return await this.findFirst({ ...options, rql: rqlWithId });
   },
 
   async findFirst(this: MailsService, options) {